+++ registry/api/v2/handler.go
@@ -74,52 +74,6 @@ type Handler struct {
 	http.Handler
 }
 
-func NewResponseWriter(w http.ResponseWriter) *ResponseWriter {
-	return &ResponseWriter{
-		ResponseWriter: w,
-		statusCode:     http.StatusOK,
-	}
-}
-
-type ResponseWriter struct {
-	http.ResponseWriter
-	statusCode    int
-	headerWritten bool
-}
-
-func (w *ResponseWriter) Write(b []byte) (int, error) {
-	w.headerWritten = true
-	return w.ResponseWriter.Write(b)
-}
-
-func (w *ResponseWriter) WriteHeader(statusCode int) {
-	w.ResponseWriter.WriteHeader(statusCode)
-
-	if !w.headerWritten {
-		w.statusCode = statusCode
-		w.headerWritten = true
-	}
-}
-
-func (w *ResponseWriter) Code() int {
-	if !w.headerWritten {
-		panic("attempt to read status code before headers have been written")
-	}
-	return w.statusCode
-}
-
-func (w *ResponseWriter) Unwrap() http.ResponseWriter {
-	return w.ResponseWriter
-}
-
-func logger(handler http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		lw := NewResponseWriter(w)
-		handler.ServeHTTP(lw, r)
-		log.Printf("%-8s %s %d %s", r.Method, r.URL.Path, lw.Code(), r.Header.Get("User-Agent"))
-	})
-}
-
 func errorHandler(w http.ResponseWriter, r *http.Request, err error) {
 	var response *ErrorResponse
 	var cerr repository.Error
